Transmit the real chunk size instead of the theoretically configured one.
In practice I do not see any reason for the server to use
the configured chunk size as it should be hardened enough to
work with any chunk size provided by the client.

On the other hand the real chunk size can be used to check
if the file transfer was successfull or not. Especially
for the last chunk that is bigger most of the time, than the configured size
we would be unable to detect if the last chunk was successfully transmitted in
case of a browser reload.