Delayed retries appear as pending with progress > 0
When using the delayed retry interval a chunk will be reset with the xhr as null.
As a result, the chunk appears as pending and is available to be sent once another
finishes.  However, if it gets sent and then the timer fires the same chunk will be
sent a second time (possibly simultaneously).

Since the loaded value isn't cleared on abort the progress for a chunk pending
retry also reports its progress as how far the chunk got before the failure.  This
can result in a chunk being 100% sent but failed on the server and the fileSuccess
firing before the chunk is actually completed.