Update on assignBrowseMethod
When using A with href='' or href='#', will often lead developers astray, since the demo, as I recall does not use href in the a tags. The following changes adds a check for A tags and preventDefaults the click. I have tested this in Firefox and seems to work fine.