Don't stall when a request times out.
Resumable allows setting a timeout on AJAX requests but doesn't handle the ontimeout event when it gets thrown. This causes the uploader to hang whenever a timeout occurs. This change will handle timeout events the same way as a load or error event is handle in the test and send methods of a chunk.