Update PHP Sample to fix recreating file before file is finished uploading
A problem was discovered in the PHP sample where the server would start
building the file before all the file’s chunks where received and
therefore corrupted the file. I have edited the code to do a check to
see if the server has the files equal to the size of the file being
uploaded before building the file.