Make sure `fileSuccess` event are not fired before the server has returned.
Do you remember installing Windows 95? How it got to 95% done and then hung for the next few hours without any progress. This commit brings the awesomeness of this to Resumable.js as well.

Before, we use the upload progress as an indicator of the full upload process -- this meant that `fileSuccess` would fire before the server had returned its status code in some cases. Since we rely on that status code, a simple padding has been made of the last 5% of the progress of each individual chunks is actually assigned to the server returning -- instead of to the pure bytes being updated. 

Given small enough chunk sizes, this should now affect the user experience of using resumable.