fixed skipping files
files which have already been uploaded (i.e. a file with the same
id already exists) should be skipped. because a non-existing
variable was used, trying to upload the same file again resulted
in a 'variable f is not defined' error.

this commit fixes the error by using the proper 'file' variable