Add support for asynchronous Blob.slice operations
The current implementation of send relies on a synchronous impl
of Blob.slice. This does not work in React Native since the
react-native-fetch-blob polyfill has an async slice method.

This PR wraps the call to `slice` in a Promise.resolve(), making it
compatible with both promises and non-promises. My only worry is
polyfilling of Promise itself. This works very well in React Native.