Add ResumableFileObject to target-callback
I needed this functionality because I had to re-calculate the target-url for each file based on information I added to the ResumableFile-Object in the "fileAdded"-Hook.

Basically the information is only available when the user adds the file to the queue. This information is not available at the time of the actual uploading. 

Therefore I add this piece of information to the ResumableFileObject when the "fileAdded"-Hook fires. Later, when the target-callback is executed, I retrieve it from the ResumableFileObject.